Use sway from NixOS module
See https://github.com/NixOS/nixpkgs/issues/128469

sway does not lunch when /run/opengl-driver provided by mesa is
different from the version it depends upon (which happens when the
system mesa != the home-manager mesa). May be that we will have to use
the same nixpkgs revision for home and system, but for now allow the
NixOS module to provide sway.