Power of "NEO"
New neovim config, based mostly on  mjlbach/defaults.nvim with some
of my old config mixed in. Unlike the old vim config, this will only
support neovim.

Has basic support for autocomplete, tree-sitter, and LSP. Most of
the bindings are DOOM Emacs inspired, but they are not a 1:1 in terms
of coverage or functionality.

Works well enough for what is currently my secondary editor, but
who knows what the future holds…